fix(math): accept fractional mode costs (2.5x) in index.json
A mode cost was parsed as u64, so any ante/buy price that is not a whole
multiple of the base bet ("cost": 2.5) failed index.json parsing outright
— stats, previews and /play all refused the revision. The protocol and the
UI already carried the cost as a decimal.

Cost is now an f64 on the wire, with the bet arithmetic running in
thousandths (u128) so bet -> stake -> base bet round-trips exactly instead
of drifting through float division. Values finer than a thousandth are
still rejected rather than silently rounding the player stake.

+++ crates/lgs/src/replay.rs
@@ -37,7 +37,7 @@ async fn replay_handler(
     let result = state.engine.replay_event(&game, &mode, event).await?;
     Ok(Json(ReplayResponse {
         payout_multiplier: result.payout_multiplier as f64 / 100.0,
-        cost_multiplier: result.cost_multiplier as f64,
+        cost_multiplier: result.cost_multiplier,
         state: result.state,
     }))
 }
